feat(button): migrate to mdx
