+++ packages/core/styles/components/button.css
@@ -1,12 +1,12 @@
 .btn {
-  @apply flex items-center gap-2 rounded py-2 px-6 text-sm font-medium leading-none;
+  @apply flex items-center gap-2 rounded py-2 px-6 text-sm font-medium leading-none bg-high-light text-primary-pure;
 
   svg,
   i {
     @apply h-4 w-4;
   }
 
-  &-md {
+  &.btn-md {
     @apply py-3 leading-normal;
     svg,
     i {
@@ -14,7 +14,7 @@
     }
   }
 
-  &-lg {
+  &.btn-lg {
     @apply py-4 text-base leading-normal;
     svg,
     i {
@@ -22,7 +22,7 @@
     }
   }
 
-  &-icon {
+  &.btn-icon {
     @apply p-2;
     svg,
     i {
@@ -46,30 +46,32 @@
     }
   }
 
-  &-solid {
+  &.btn-primary {
+    @apply bg-primary-pure text-white hover:bg-primary-medium focus:outline-none focus:ring-2 focus:ring-primary-lighter active:bg-primary-dark disabled:bg-high-medium disabled:text-high-dark;
+  }
+  &.btn-secondary {
+    @apply bg-secondary-pure text-white hover:bg-secondary-medium focus:outline-none focus:ring-2 focus:ring-secondary-light active:bg-secondary-dark disabled:bg-high-medium disabled:text-high-dark;
+  }
+
+  &.btn-outline {
+    @apply bg-transparent border focus:outline-none focus:ring-2;
+
     &.btn-primary {
-      @apply bg-primary-pure text-white hover:bg-primary-medium focus:outline-none focus:ring-2 focus:ring-primary-lighter active:bg-primary-dark disabled:bg-high-medium disabled:text-high-dark;
+      @apply border-primary-pure hover:text-white text-primary-pure hover:bg-primary-medium focus:ring-primary-lighter active:bg-primary-dark disabled:border-high-dark disabled:bg-transparent disabled:text-high-dark;
     }
     &.btn-secondary {
-      @apply bg-secondary-pure text-white hover:bg-secondary-medium focus:outline-none focus:ring-2 focus:ring-secondary-light active:bg-secondary-dark disabled:bg-high-medium disabled:text-high-dark;
+      @apply border-secondary-pure hover:text-white text-secondary-pure hover:bg-secondary-medium focus:ring-secondary-light active:bg-secondary-dark disabled:border-high-dark disabled:bg-transparent disabled:text-high-dark;
     }
   }
 
-  &-outline {
-    &.btn-primary {
-      @apply border border-primary-pure text-primary-pure hover:bg-primary-medium hover:text-white focus:outline-none focus:ring-2 focus:ring-primary-lighter active:bg-primary-dark disabled:border-high-dark disabled:bg-transparent disabled:text-high-dark;
-    }
-    &.btn-secondary {
-      @apply border border-secondary-pure text-secondary-pure hover:bg-secondary-medium hover:text-white focus:outline-none focus:ring-2 focus:ring-secondary-light active:bg-secondary-dark disabled:border-high-dark disabled:bg-transparent disabled:text-high-dark;
-    }
-  }
+  &.btn-ghost {
+    @apply bg-transparent focus:outline-none focus:ring-2;
 
-  &-ghost {
     &.btn-primary {
-      @apply text-primary-pure hover:bg-high-light focus:outline-none focus:ring-2 focus:ring-primary-lighter active:bg-high-dark disabled:bg-transparent disabled:text-high-dark;
+      @apply text-primary-pure hover:bg-high-light focus:ring-primary-lighter active:bg-high-dark disabled:bg-transparent disabled:text-high-dark;
     }
     &.btn-secondary {
-      @apply text-secondary-pure hover:bg-high-light focus:outline-none focus:ring-2 focus:ring-secondary-light active:bg-high-dark disabled:bg-transparent disabled:text-high-dark;
+      @apply text-secondary-pure hover:bg-high-light focus:ring-secondary-light active:bg-high-dark disabled:bg-transparent disabled:text-high-dark;
     }
   }
 }
