+++ storybook/stories/components/avatar/Avatar.stories.mdx
@@ -10,7 +10,7 @@ export const Template = args => <Avatar {...args} />;
 
 ## Default
 
-<div className="pt-3">
+<div className="pt-4">
   <Avatar />
 </div>
 
@@ -18,7 +18,7 @@ export const Template = args => <Avatar {...args} />;
 
 ## Colors
 
-<div className="flex items-center gap-3 pt-3">
+<div className="flex items-center gap-3 pt-4">
   <Avatar />
   <Avatar color="primary" />
 </div>
@@ -34,7 +34,7 @@ export const Template = args => <Avatar {...args} />;
 
 ## Sizes
 
-<div className="flex items-start gap-3 pt-3">
+<div className="flex items-start gap-3 pt-4">
   <Avatar size="small" />
   <Avatar />
   <Avatar size="large" />
@@ -52,7 +52,7 @@ export const Template = args => <Avatar {...args} />;
 
 ## Active State
 
-<div className="flex items-start gap-3 pt-3">
+<div className="flex items-start gap-3 pt-4">
   <Avatar isActive />
   <Avatar isActive color="primary" />
 </div>
@@ -68,7 +68,7 @@ export const Template = args => <Avatar {...args} />;
 
 ## Content Types
 
-<div className="flex items-start gap-3 pt-3">
+<div className="flex items-start gap-3 pt-4">
   <Avatar content="text" />
   <Avatar content="icon" />
   <Avatar content="image" />
