+++ storybook/stories/components/badge/Badge.stories.mdx
@@ -9,7 +9,7 @@ Badges are used to highlight important information.
 
 ## Default
 
-<div className="pt-3">
+<div className="pt-4">
   <span className="badge">32</span>
 </div>
 
@@ -23,7 +23,7 @@ Badges are used to highlight important information.
 
 ## Colors
 
-<div className="flex items-center gap-3 pt-3">
+<div className="flex items-center gap-3 pt-4">
   <span className="badge">32</span>
   <span className="badge badge-highlight">32</span>
 </div>
