+++ storybook/stories/components/breadcrumb/Breadcrumb.stories.mdx
@@ -7,7 +7,7 @@ import { Meta, Source } from '@storybook/addon-docs';
 
 ## Default
 
-<nav className="pt-3" aria-label="breadcrumb">
+<nav className="pt-4" aria-label="breadcrumb">
   <ol>
     <li>
       <a href="">Home</a>
