--- storybook/stories/components/button/Button.ts
@@ -1,52 +1,0 @@
-import clsx from 'clsx';
-
-export type ButtonArgs = {
-  size?: 'small' | 'medium' | 'large';
-  variant?: 'unstyled' | 'solid' | 'outline' | 'ghost';
-  color?: 'unstyled' | 'primary' | 'secondary';
-  withIcon?: boolean;
-  isDisabled?: boolean;
-};
-
-const sizes = {
-  small: '',
-  medium: 'btn-md',
-  large: 'btn-lg',
-};
-
-const variants = {
-  unstyled: '',
-  solid: 'btn-solid',
-  outline: 'btn-outline',
-  ghost: 'btn-ghost',
-};
-
-const colors = {
-  unstyled: '',
-  primary: 'btn-primary',
-  secondary: 'btn-secondary',
-};
-
-export function createButton({
-  size = 'small',
-  color = 'unstyled',
-  variant = 'unstyled',
-  isDisabled = false,
-  withIcon = false,
-}: ButtonArgs) {
-  const button = document.createElement('button');
-  button.type = 'button';
-  button.className = clsx('btn', sizes[size], colors[color], variants[variant]);
-
-  if (isDisabled) button.disabled = true;
-  if (withIcon) {
-    button.innerHTML = `
-    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor">
-      <path stroke-linecap="round" stroke-linejoin="round" d="M15.75 10.5V6a3.75 3.75 0 10-7.5 0v4.5m11.356-1.993l1.263 12c.07.665-.45 1.243-1.119 1.243H4.25a1.125 1.125 0 01-1.12-1.243l1.264-12A1.125 1.125 0 015.513 7.5h12.974c.576 0 1.059.435 1.119 1.007zM8.625 10.5a.375.375 0 11-.75 0 .375.375 0 01.75 0zm7.5 0a.375.375 0 11-.75 0 .375.375 0 01.75 0z" />
-    </svg> Shop now`;
-  } else {
-    button.textContent = 'Shop now';
-  }
-
-  return button;
-}
